workshop
# workshop
lab exercise
1.write a program that takes three number and print their sum.every number is given in seperate line

2.write a pogram that reads the length of base and height of right andle triangle and print the area.
every numberis given on a seperate line

3.n students takes k apples and distribute among them evenly the remaining (the individial) part remain in the basket
3.how many apples will each single student get? how many apple ramin in the basket ?the program reads the number N and K
it should print two anser for the question above
given the integer N -the number of minute that is passed since midnigth -

4.how many hours and minute are displayed  on 24h digital clock?
the program should print two numbers: the number of hours (between 0  and 25)and the number of minute ( between 0 and 59)
a school decided to replace the desks in three classroom . each desk sits two students given the number of
students in each class.
print the smallest possible number of desks that can be purchased the program should read three integers the
number of students in each of the three classes a,b and c respectively.
In the first test there is three groups .the first group has 20 students and thus needs 10 desks. the second
group has 21 students , so they can get by with no fewer than 11 desks . 11 desks are also enough for third group
of 22 students.so, we need 32 desks in total .  

5.solve the each of the following problem using python scripts.make sure you use an appropriate variable names and comments.
when there is final answer have python print it to the screen

6.A person's body mass BMI id defined as:
 BMI : mass in kg /(height in meter)2
 
7.you live 4 miles from university .  the bus driver drives at 25mph but spends 2 minute at each of the ten stops on the
way .how long will the bus journey takes ? .alternatively , you could run to university. you jog the first mile  at 7 mph
then run the next 2 at 15mph before jogging the last at 7mph again . will this be quicker or slower than bus ?

8.write the python program which  accepts  the radious of a circle from the user and complete  the area .
area of circle = pi * r **2

9.write the python program to find sum of first  n positive number
sum = (n*(n+1))/2

10.write a python program to  convert seconds to day ,hour ,minute and second